requirements.txt
alembic==1.7.4
argon2-cffi==21.1.0
asgiref==3.4.1
attrs==21.2.0
black==21.9b0
blake3==0.3.1
boto3==1.18.59
botocore==1.21.59
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.2
configparser==5.0.2
cryptography==35.0.0
dnspython==2.1.0
email-validator==1.1.3
falcon==3.0.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
hashids==1.3.1
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.2.2
iniconfig==1.1.1
jmespath==0.10.0
Mako==1.1.5
MarkupSafe==2.0.1
mypy-extensions==0.4.3
numpy==1.22.1
packaging==21.0
pandas==1.3.5
passlib==1.7.4
pathspec==0.9.0
Pillow==9.0.0
pkg_resources==0.0.0
platformdirs==2.4.0
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.10.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.2.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
regex==2021.10.8
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.25
stripe==2.61.0
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
twilio==7.5.0
typing-extensions==3.10.0.2
urllib3==1.26.7
zipp==3.6.0
